Fix the bridge direction and a bit of elephant foot compensation.
The fork works well with 10mm spacing and sounds like the real thing! Well, except that it vibrates as 135Hz, almost exactly halfway between C3 and C#3. Go figure - the problem is somewhere in the stiffness calculation, i.e. I or A are wrong.

OTOH, the ABS C4 fork is SPOT ON at C4, 262Hz.

The PLA C4 fork is a slightly flat B3 at 245Hz.